notifikasi

+++ database/migrations/2025_06_21_120917_add_koin_dibutuhkan_to_rewards_table.php
@@ -7,16 +7,20 @@ use Illuminate\Support\Facades\Schema;
 class AddKoinDibutuhkanToRewardsTable extends Migration
 {
     public function up()
-    {
+{
+    if (Schema::hasTable('rewards')) {
         Schema::table('rewards', function (Blueprint $table) {
             $table->integer('koin_dibutuhkan')->nullable()->after('subtipe_gacha');
         });
     }
+}
 
-    public function down()
-    {
+public function down()
+{
+    if (Schema::hasTable('rewards')) {
         Schema::table('rewards', function (Blueprint $table) {
             $table->dropColumn('koin_dibutuhkan');
         });
     }
+}
 }
